+++ bin/check-dns.rb
@@ -82,6 +82,21 @@ class DNS < Sensu::Plugin::Check::CLI
     entries
   end
 
+  def check_against_regex(entries, regex)
+    # produce an Array of entry strings
+    b = if entries.answer.count > 1
+          entries.answer.rrsets(config[:type].to_s).map(&:to_s)
+        else
+          [entries.answer.first.to_s]
+        end
+    b.each do |answer|
+      if answer.match(regex)
+        ok "Resolved #{config[:domain]} #{config[:type]} matched #{regex}"
+      end
+    end # b.each()
+    critical "Resolved #{config[:domain]} #{config[:type]} did not match #{regex}"
+  end
+
   def run
     unknown 'No domain specified' if config[:domain].nil?
 
@@ -112,18 +127,7 @@ class DNS < Sensu::Plugin::Check::CLI
         critical "Resolved #{config[:domain]} #{config[:type]} did not include #{config[:result]}"
       end
     elsif config[:regex]
-      # produce an Array of entry strings
-      b = if entries.answer.count > 1
-            entries.answer.rrsets(config[:type].to_s).map(&:to_s)
-          else
-            [entries.answer.first.to_s]
-          end
-      b.each do |answer|
-        if answer.match(Regexp.new(config[:regex]))
-          ok "Resolved #{config[:domain]} #{config[:type]} matched #{config[:regex]}"
-        end
-      end # b.each()
-      critical "Resolved #{config[:domain]} #{config[:type]} did not match #{config[:regex]}"
+      check_against_regex(entries, Regexp.new(config[:regex]))
     else
       ok "Resolved #{config[:domain]} #{config[:type]}"
     end