+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ## [Unreleased]
 ### Added
 * Adding support for Regular Expression matching
+* Added simple response time metric
 
 ## [0.0.6] - 2015-09-28
 ### Changed